requirements.txt
streamlit==1.15.0
openai==0.27.0
duckduckgo-search==0.5.0
biopython==1.81
requests==2.28.1
pillow==9.3.0
pandas==1.5.2
numpy==1.23.3
